Tested on actual hardware. Works great.
