+++ test/pubspec.yaml
@@ -1,4 +1,3 @@
 name: tester
 dependencies:
   async: any
-  test: any
