+++ test/readme.txt
@@ -1,0 +1,8 @@
+First get the packages:
+
+* pub get
+
+Then run the tests
+
+* dart test.dart
+
